+++ src/schema/resolvers/user.js
@@ -5,16 +5,29 @@ require('dotenv').config();
 
 module.exports = {
   Query: {
-    me: async (parent, { user }) => {
+    me: async (parent, args, { user }) => {
+      console.log(user);
       if (!user) {
         throw new Error('You are not authenticated!');
       }
 
+      console.log(user);
+
       // user is authenticated
       return await User.findById(user.id);
     },
-    user: async (parent, { id }) => await User.findById(id),
-    users: async () => await User.find({})
+    user: async (parent, { id }, { user }) => {
+      if (!user.id === id && !user.admin) {
+        throw new Error('Not authorized');
+      }
+      return await User.findById(id);
+    },
+    users: async (parent, args, { user }) => {
+      if (!user.admin) {
+        throw new Error('Not authorized');
+      }
+      return await User.find({});
+    }
   },
 
   Mutation: {
@@ -42,13 +55,38 @@ module.exports = {
         { expiresIn: '1d' }
       );
     },
-    removeUser: async (parent, { id }) => {
+    updateUser: async (parent, args, { user }) => {
+      const { id, admin, ...updatedProperties } = args;
+
+      if (!user.id === id && !user.admin) {
+        throw new Error('Not authorized');
+      }
+
+      if (user.admin) {
+        updatedProperties.admin = admin;
+      }
+
+      const updatedUser = await User.findByIdAndUpdate(id, {
+        $set: { ...updatedProperties }
+      });
+
+      if (!updatedUser) {
+        throw new Error('User not found');
+      }
+
+      return updatedUser;
+    },
+    removeUser: async (parent, { id }, { user }) => {
+      if (!user.id === id && !user.admin) {
+        throw new Error('Not authorized');
+      }
       const removedUser = await User.findByIdAndRemove(id);
+
       if (!removedUser) {
         throw new Error('User not found');
       }
 
-      return removedUser;
+      return removedUser._id;
     }
   }
 };
